+++ src/commands.rs
@@ -2,7 +2,7 @@ use crate::Cli;
 use crate::Commands;
 use chrono::NaiveTime;
 use r_timelog::config::Config;
-use r_timelog::utils::mins2hhmm;
+use r_timelog::utils::{mins2hhmm, print_separator};
 use r_timelog::{db, logic, utils};
 use rusqlite::Connection;
 use std::process::Command;
@@ -162,9 +162,15 @@ pub fn handle_add(cmd: &Commands, db_path: &str) -> rusqlite::Result<()> {
 }
 
 /// Handle the `list` command
-pub fn handle_list(period: Option<String>, db_path: &str) -> rusqlite::Result<()> {
+pub fn handle_list(
+    period: Option<String>,
+    pos: Option<String>,
+    db_path: &str,
+) -> rusqlite::Result<()> {
     let conn = Connection::open(db_path)?;
-    let sessions = db::list_sessions(&conn, period.as_deref())?;
+    // Normalize pos to uppercase
+    let pos_upper = pos.as_ref().map(|p| p.trim().to_uppercase());
+    let sessions = db::list_sessions(&conn, period.as_deref(), pos_upper.as_deref())?;
 
     if sessions.is_empty() {
         println!("⚠️  No recorded sessions found");
@@ -184,13 +190,16 @@ pub fn handle_list(period: Option<String>, db_path: &str) -> rusqlite::Result<()
                 year
             );
         }
+    } else if let Some(p) = pos.as_deref() {
+        println!("📅 Saved sessions for position {}:", p);
     } else {
         println!("📅 Saved sessions:");
     }
 
+    let mut total_surplus = 0;
+
     for s in sessions {
         if s.position == "H" {
-            //println!("{:>3}: {} | \x1b[35mHoliday\x1b[0m",s.id,s.date);
             println!(
                 "{:>3}: {} | \x1b[1;37;45m {:29}Holiday{:30} \x1b[0m",
                 s.id, s.date, "", ""
@@ -205,19 +214,41 @@ pub fn handle_list(period: Option<String>, db_path: &str) -> rusqlite::Result<()
         if has_start && !has_end {
             // Only start → calculate expected end
             let expected = logic::calculate_expected_exit(&s.start, work_minutes, s.lunch);
-            // \x1b[90mLunch {}\x1b[0m
+
+            let lunch_color = if s.lunch > 0 { "\x1b[0m" } else { "\x1b[90m" };
+            let lunch_str = if s.lunch > 0 {
+                mins2hhmm(s.lunch)
+            } else {
+                "-".to_string()
+            };
+            // Forza la larghezza a 5 caratteri, allineato a destra
+            let lunch_fmt = format!("{:^5}", lunch_str);
+
+            let end_color = if !s.end.is_empty() {
+                "\x1b[0m"
+            } else {
+                "\x1b[90m"
+            };
+            let end_str = if !s.end.is_empty() {
+                s.end
+            } else {
+                "-".to_string()
+            };
+            // Forza la larghezza a 5 caratteri, allineato a destra
+            let end_fmt = format!("{:^5}", end_str);
+
             println!(
-                "{:>3}: {} | Position {} | Start {} | Lunch {} | \x1b[90mEnd   -\x1b[0m   | Expected {} | \x1b[90mSurplus   -\x1b[0m",
+                "{:>3}: {} | Position {} | Start {} | {}Lunch {}\x1b[0m | {}End {}\x1b[0m | Expected {} | \x1b[90mSurplus {:^8}\x1b[0m",
                 s.id,
                 s.date,
                 s.position,
                 s.start,
-                if s.lunch > 0 {
-                    mins2hhmm(s.lunch)
-                } else {
-                    "-".to_string()
-                },
+                lunch_color,
+                lunch_fmt,
+                end_color,
+                end_fmt,
                 expected.format("%H:%M"),
+                "-",
             );
         } else if has_start && has_end {
             let start_time = NaiveTime::parse_from_str(&s.start, "%H:%M").unwrap();
@@ -236,6 +267,7 @@ pub fn handle_list(period: Option<String>, db_path: &str) -> rusqlite::Result<()
                 let surplus =
                     logic::calculate_surplus(&s.start, effective_lunch, &s.end, work_minutes);
                 let surplus_minutes = surplus.num_minutes();
+                total_surplus += surplus_minutes;
 
                 let color_code = if surplus_minutes < 0 {
                     "\x1b[31m"
@@ -251,17 +283,22 @@ pub fn handle_list(period: Option<String>, db_path: &str) -> rusqlite::Result<()
                     format!("{:+}", surplus_minutes)
                 };
 
+                let lunch_str = if effective_lunch > 0 {
+                    mins2hhmm(effective_lunch)
+                } else {
+                    "-".to_string()
+                };
+
+                // Forza la larghezza a 5 caratteri, allineato a destra
+                let lunch_fmt = format!("{:^5}", lunch_str);
+
                 println!(
                     "{:>3}: {} | Position {} | Start {} | Lunch {} | End {} | Expected {} | Surplus {}{:>4} min\x1b[0m",
                     s.id,
                     s.date,
                     s.position,
                     s.start,
-                    if effective_lunch > 0 {
-                        mins2hhmm(effective_lunch)
-                    } else {
-                        "-".to_string()
-                    },
+                    lunch_fmt,
                     s.end,
                     expected.format("%H:%M"),
                     color_code,
@@ -270,33 +307,67 @@ pub fn handle_list(period: Option<String>, db_path: &str) -> rusqlite::Result<()
             } else {
                 // Case without lunch (work entirely outside the window)
                 let duration = end_time - start_time;
+                let lunch_str = "-".to_string();
+
+                // Forza la larghezza a 5 caratteri, allineato a destra
+                let lunch_fmt = format!("{:^5}", lunch_str);
+
                 println!(
-                    "{:>3}: {} | Position {} | Start {} | \x1b[90mLunch   -\x1b[0m   | End {} | \x1b[36mWorked {:>2} h {:02} min\x1b[0m",
+                    "{:>3}: {} | Position {} | Start {} | \x1b[90mLunch {}\x1b[0m | End {} | \x1b[36mWorked {:>2} h {:02} min\x1b[0m",
                     s.id,
                     s.date,
                     s.position,
                     s.start,
+                    lunch_fmt,
                     s.end,
                     duration.num_hours(),
                     duration.num_minutes() % 60
                 );
             }
         } else {
+            let lunch_str = if s.lunch > 0 {
+                mins2hhmm(s.lunch)
+            } else {
+                "-".to_string()
+            };
+
+            // Forza la larghezza a 5 caratteri, allineato a destra
+            let lunch_fmt = format!("{:^5}", lunch_str);
+
             // Incomplete information
             println!(
-                "{:>3}: {} | Position {} | Start {} | Lunch {} | End {}",
+                "{:>3}: {} | Position {} | Start {:^5} | Lunch {} | End {:^5}",
                 s.id,
                 s.date,
                 s.position,
                 if has_start { &s.start } else { "-" },
-                if s.lunch > 0 {
-                    mins2hhmm(s.lunch)
-                } else {
-                    "-".to_string()
-                },
+                lunch_fmt,
                 if has_end { &s.end } else { "-" }
             );
         }
+    }
+
+    println!();
+    print_separator('-', 25, 104);
+
+    if total_surplus != 0 {
+        let color_code = if total_surplus < 0 {
+            "\x1b[31m" // rosso
+        } else {
+            "\x1b[32m" // verde
+        };
+
+        let formatted_total = format!("{:+}", total_surplus);
+
+        println!(
+            "{:>113}",
+            format!(
+                "Σ Total surplus: {}{:>4} min\x1b[0m",
+                color_code, formatted_total
+            ),
+        );
+    } else {
+        println!("{:>113}", format!("Σ Total surplus: {:>4} min", 0));
     }
 
     Ok(())